Store one-bit flags as bool bitfields
`unsigned IsFoo : 1` flags become `bool IsFoo : 1` across both games and
the shared code (597 declarations, including the infantry DoStruct
tables and the options flags), and assignments of 0/1 to them become
false/true. Saves already copy bitfields through bool temporaries, so
the archive format is unchanged.

Layout-sensitive types keep `unsigned`: SpecialClass, EventClass, the
anonymous packet structs and the cell flag unions. Multi-bit fields that
now sit next to bool flags (flash count, harvested ore, nuke pieces,
radar frames, TD targetting mode) use uint8_t storage so the Microsoft
ABI still packs them. TD's recording file stores GameOptionsClass raw,
so recordings from older builds no longer load.

The TD IPX manager sets its members in the initializer list, with the
two state flags defaulted in the class.

+++ src/ra/techno.cc
@@ -2499,7 +2499,7 @@ HousesType TechnoClass::Owner() const {
 void TechnoClass::Clicked_As_Target(int count) {
   assert(IsActive);
 
-  FlashCount = static_cast<unsigned>(count);
+  FlashCount = static_cast<uint8_t>(count);
 }
 
 /***********************************************************************************************
